Player objective interaction

+++ common/game/Evolve.h
@@ -3,14 +3,15 @@
 
 #include "Objective.h"
 
+
 class Evolve : public Objective {
     private:
-        int evo_amount; // amount to evolve
+        float evo_amount; // amount to evolve
 
     public:
         Evolve(GridPosition position, Restriction r = R_MONSTER);
 
-        int getEvoAmount();
+        float getEvoAmount();
 };
 
 #endif
